RavenDB-13556 create a seperate class for the rachis history log
set log entries as cancel if the follower need to rewind
